+++ src/corpus/corpus.composable.ts
@@ -102,14 +102,7 @@ export function useCorpus(corpusId: string) {
   async function downloadSource(source: FileMeta, binary: boolean) {
     return spin(
       api.downloadSources(corpusId, source.name, binary),
-      `corpus/${corpusId}/sources/${source.name}/raw`,
-    ).catch(alertError);
-  }
-
-  async function downloadPlaintext(source: FileMeta) {
-    return spin(
-      api.downloadSourceText(corpusId, source.name),
-      `corpus/${corpusId}/sources/${source.name}/plain`,
+      `corpus/${corpusId}/sources/${source.name}`,
     ).catch(alertError);
   }
 
@@ -176,7 +169,6 @@ export function useCorpus(corpusId: string) {
     sources,
     hasSources,
     downloadSource,
-    downloadPlaintext,
     uploadSources,
     deleteSource,
     extensions,